Fixed unexplained PC shutdown. (Win10 driver bug?)
It appears that the Win10 UPS driver will occasionally copy whatever
the present RemainingCapacity is to the RemCapacityLimit; sending it
to the Leonardo. We now report when it happens, but ignore the setting.
Also moved many strings (for printouts) to flash: F(""), freeing up RAM.
Added ProjectDefs.h which now holds SERIAL1_DEBUG and related.

I tested with Ubuntu on a laptop, and in CDC Enabled mode, it did not show up,
but in CDC Disabled mode it did show up. Ubuntu never seemed to send the
RemCapacityLimit setting to the Leo (yay), but my debug prints indicate
that the Leo is not successfully sending the PowerDevice.sendReport's.
These printed out to Serial1:
     Sending: Batt Remaining = 88%, 105:36, Discharging: N, Status = 0x8
     Comms with PC (neg=bad): -1,-1
Ubuntu continuted to show the initial Leo "battery" capacity, and it never
updated. On Win10 it did eventually update.