Fixed spelling autocorrect in VersionNotes.txt
And deleted MergeNotes_Sid.txt as it was outdated and is functionally replaced by VersionNotes.txt:
Fixed unexplained PC shutdown. (Win10 driver bug?)
It appears that the Win10 UPS driver will occasionally sent whatever
the present HID_PD_REMAININGCAPACITY is to the HID_PD_REMNCAPACITYLIMIT; sending it
to the Leonardo. We now report when that problem happens, but ignore the setting.
Also moved many strings (for printouts) to flash: F(""), freeing up RAM.
Added projected.h which now holds SERIAL1_DEBUG and related.

I tested with Ubuntu on a laptop, and in CDC Enabled mode, it did not show up,
but in CDC Disabled mode it did show up. Ubuntu never seemed to send the
HID_PD_REMNCAPACITYLIMIT setting to the Leo (yay), but my debug prints indicate
that the Leo is not successfully sending the PowerDevice.sendReport's.
These printed out to Serial1:
     Sending: Batt Remaining = 88%, 105:36, Discharging: N, Status = 0x8
     Comms with PC (neg=bad): -1,-1
Ubuntu continued to show the initial Leo "battery" capacity, and it never
updated. On Win10 it did eventually update.